Refactor time handling in activity and participant components to support milliseconds
- Updated the activity update function to reference event_type instead of event_id for validation.
- Changed the time field in the Activity model from float to int for better precision.
- Modified frontend components to convert time from milliseconds to a formatted string.
- Enhanced time parsing functions to support various formats, including milliseconds.
- Updated UI components to ensure consistent time formatting and improved user experience.

+++ backend/utils/variables.py
@@ -23,7 +23,7 @@ class Activity(BaseModel):
     weight: float | None = None
     type_of_activity: str
     reps: int | None = None
-    time: float | None = None
+    time: int | None = None
     is_success: bool | None = None
     is_deleted: bool | None = None
     event_type: int | None = None
